+++ modules/Students/includes/AddressFields.php
@@ -82,7 +82,10 @@ if ( isset( $_POST['tables'] )
 					{
 						if ( $table === 'address_fields' )
 						{
-							AddDBField( 'address', $id, $columns['TYPE'] );
+							// @since 12.8 SQL add comment to Field column
+							$comment = ParseMLField( $columns['TITLE'] );
+
+							AddDBField( 'address', $id, $columns['TYPE'], $comment );
 
 							$_REQUEST['id'] = $id;
 						}