+++ modules/Students/includes/StudentFields.php
@@ -104,7 +104,10 @@ if ( isset( $_POST['tables'] )
 					{
 						if ( $table === 'custom_fields' )
 						{
-							AddDBField( 'students', $id, $columns['TYPE'] );
+							// @since 12.8 SQL add comment to Field column
+							$comment = ParseMLField( $columns['TITLE'] );
+
+							AddDBField( 'students', $id, $columns['TYPE'], $comment );
 
 							$_REQUEST['id'] = $id;
 						}